Move window.Window() up one level from aurora/ to processing/
