+++ mt_metadata/transfer_functions/processing/__init__.py
@@ -1,1 +1,2 @@
 from .time_series_decimation import TimeSeriesDecimation
+from .window import Window
