+++ mt_metadata/transfer_functions/processing/standards/window.json
