+++ mt_metadata/transfer_functions/processing/window.py
